Begins split out of snapshots from the main FSM class.
